Move blob schedule test & delete some invariant tests
