+++ tests/core/pyspec/eth2spec/test/deneb/unittests/test_config_invariants.py
@@ -1,34 +1,19 @@
 from eth2spec.test.context import (
     single_phase,
-    spec_state_test,
+    spec_test,
     with_deneb_and_later,
+    with_presets,
+)
+from eth2spec.test.helpers.constants import (
+    MAINNET,
 )
 from eth2spec.test.helpers.forks import is_post_eip7732
 
 
 @with_deneb_and_later
-@spec_state_test
+@spec_test
 @single_phase
-def test_length(spec, state):
-    assert (
-        spec.get_max_blobs_per_block(spec.get_current_epoch(state))
-        < spec.MAX_BLOB_COMMITMENTS_PER_BLOCK
-    )
-
-
-@with_deneb_and_later
-@spec_state_test
-@single_phase
-def test_networking(spec, state):
-    assert (
-        spec.config.MAX_REQUEST_BLOB_SIDECARS
-        == spec.config.MAX_REQUEST_BLOCKS_DENEB
-        * spec.get_max_blobs_per_block(spec.get_current_epoch(state))
-    )
-    # Start with the same size, but `BLOB_SIDECAR_SUBNET_COUNT` could potentially increase later.
-    assert spec.config.BLOB_SIDECAR_SUBNET_COUNT == spec.get_max_blobs_per_block(
-        spec.get_current_epoch(state)
-    )
+def test_networking(spec):
     for i in range(spec.MAX_BLOB_COMMITMENTS_PER_BLOCK):
         if is_post_eip7732(spec):
             inner_gindex = spec.get_generalized_index(
@@ -45,3 +30,15 @@ def test_networking(spec, state):
         else:
             gindex = spec.get_generalized_index(spec.BeaconBlockBody, "blob_kzg_commitments", i)
             assert spec.floorlog2(gindex) == spec.KZG_COMMITMENT_INCLUSION_PROOF_DEPTH
+
+
+@with_deneb_and_later
+@spec_test
+@single_phase
+@with_presets([MAINNET], reason="to have fork epoch number")
+def test_blob_schedule(spec):
+    for entry in spec.config.BLOB_SCHEDULE:
+        # Check that all epochs are post-Deneb
+        assert entry["EPOCH"] >= spec.config.DENEB_FORK_EPOCH
+        # Check that all blob counts are less than the limit
+        assert entry["MAX_BLOBS_PER_BLOCK"] <= spec.MAX_BLOB_COMMITMENTS_PER_BLOCK
