+++ tests/core/pyspec/eth2spec/test/electra/unittests/test_config_invariants.py
@@ -1,7 +1,6 @@
 from eth2spec.test.context import (
     single_phase,
     spec_test,
-    spec_state_test,
     with_electra_and_later,
 )
 
@@ -11,22 +10,3 @@ from eth2spec.test.context import (
 @single_phase
 def test_processing_pending_partial_withdrawals(spec):
     assert spec.MAX_PENDING_PARTIALS_PER_WITHDRAWALS_SWEEP < spec.MAX_WITHDRAWALS_PER_PAYLOAD
-
-
-@with_electra_and_later
-@spec_state_test
-@single_phase
-def test_networking(spec, state):
-    assert (
-        spec.get_max_blobs_per_block(spec.get_current_epoch(state))
-        <= spec.MAX_BLOB_COMMITMENTS_PER_BLOCK
-    )
-    assert (
-        spec.config.MAX_REQUEST_BLOB_SIDECARS_ELECTRA
-        == spec.config.MAX_REQUEST_BLOCKS_DENEB
-        * spec.get_max_blobs_per_block(spec.get_current_epoch(state))
-    )
-    # Start with the same size, but `BLOB_SIDECAR_SUBNET_COUNT` could potentially increase later.
-    assert spec.config.BLOB_SIDECAR_SUBNET_COUNT_ELECTRA == spec.get_max_blobs_per_block(
-        spec.get_current_epoch(state)
-    )
