+++ tests/core/pyspec/eth2spec/test/fulu/unittests/test_config_invariants.py
@@ -1,11 +1,7 @@
 from eth2spec.test.context import (
     single_phase,
     spec_test,
-    with_presets,
     with_fulu_and_later,
-)
-from eth2spec.test.helpers.constants import (
-    MAINNET,
 )
 
 
@@ -29,15 +25,3 @@ def test_invariants(spec):
 @single_phase
 def test_polynomial_commitments_sampling(spec):
     assert spec.FIELD_ELEMENTS_PER_EXT_BLOB == 2 * spec.FIELD_ELEMENTS_PER_BLOB
-
-
-@with_fulu_and_later
-@spec_test
-@single_phase
-@with_presets([MAINNET], reason="to have fork epoch number")
-def test_blob_schedule(spec):
-    for entry in spec.config.BLOB_SCHEDULE:
-        # Check that all epochs are post-Deneb
-        assert entry["EPOCH"] >= spec.config.DENEB_FORK_EPOCH
-        # Check that all blob counts are less than the limit
-        assert entry["MAX_BLOBS_PER_BLOCK"] <= spec.MAX_BLOB_COMMITMENTS_PER_BLOCK
